requirements.txt
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Flask-WTF==1.2.1
email_validator==2.2.0
python-dotenv==1.0.1
Werkzeug==3.0.3
gunicorn==22.0.0
